+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v2.11.4
 
-
+* added `int`, `str` and `float` to `Predicate`