+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v2.11.4
 
 * added `int`, `str` and `float` to `Predicate`
+* added function overloading